Fix test of AABB_tree (use of an epsilon)
The test 'aabb_naive_vs_tree_distance_triangle_test' failed
something. It was due to a test using an epsilon, with a too-low epsilon
value (1e-8). I managed to reproduce the bug, and the quantity compared
to the epsilon value was about 1.8e-8. I increase the epsilon value to
1e-7, and add a few lines that show the compared values, when the test
fails.

Trivial bug-fix.