Fix a bug in Mesh_3
Benjamin Kehlet reported on CGAL-discuss (Tue, 25 Mar 2014 13:45:24
+0100) that Mesh_3 was not able to mesh the simple following OFF:

OFF
4 4 0

0 0 0
0 0 1
0 1 0
1 0 0
3  3 1 2
3  0 1 3
3  0 3 2
3  0 2 1

The unit tetrahedron!!

The bug was in the robust weighted circumcenter traits. The traits
checks that the circumcenter of four points is at least in the circum
sphere of the four points. But for weighted points, that is
"power_test" that must be tested, instead of "side_of_oriented_sphere".