Remove some bad assert in test_RT_cell_base_with_weighted_circumcenter_3.cpp.
These asserts should have tested that the wighted_circumcenter is well
invalidate in some cases. However, we can't access to the private pointer.
So we can't check that the pointer is NULL when invalidate_circumcenter is
called. These asserts tested that the new computed weighted circumcenter has
a different address after invalidate_circumcenter. But, that fact isn't
guaranteed.