Created Qt Demo for Polyhedron_shortest_path
A shortest paths object is created using 'Make Shortest Path' from the
menu on a polyhedron.
Source points are created by shift-clicking on the polyhedron.
Points can also be removed by selecting the appropriate option in the
combo box.  Point locations can be snapped to the nearest edge or vertex,
or placed anywhere on the face.  Choosing 'Shortest Path' from the combo
box and shift-clicking will create a polyline object representing the
shortest path from any of the source points to that destination.  Note
that computing the shortest paths tree may be slow on the first query.