Fix a bug when LEDA is used instead of GMP
That is a really trivial bug fix, for a rather nasty bug! I spend
several hours on it! Currently the `leda::rational::operator>>` does not
clear the `failbit` of the stream, when it reads a number without
denominator, and the EOF directly after. The flag `eofbit` and `failbit`
are set while trying to read the '/', but in that case the `failbit`
should be cleared from the stream status.

Trivial bug fix for master.