use a tag to choose bisector constructions
The SDG traits define a tag named Tag_has_bisector_constructions
that is either Tag_false or Tag_true.

If it is Tag_false, the bisector constructions are the hard-coded
ones from the algorithm (essentially they are the L2 ones).

If it is Tag_true, the bisector constructions are taken from the
traits. This is how the Linf traits are implemented.

This way, we avoid SFINAE.