bug fix in restoration of Cell data after revert_move
we still don't backup infinite cells because the infinite vertex does not have
a proper vertex ID (i.e. meshing_info())
However the facet data should be restored in Delaunay-restricted facets that
belong to an infinite cell (i.e. mirror facets of restricted facets that
belong to the convex hull)