Bug-fix in Mesh_3 with time stamps
The implementation of the class template Protect_edges_sizing_field was
not correct when time stamps are used for Vertex_handle. The issue is
that the time stamp of a vertex change when one do:
  tr.remove(v);
  v = tr.insert(point);
The vertex pointed by 'v' is re-used by the TDS, but the time stamp of
the vertex does change.