Fix a warning -Wmaybe-uninitialized
Workaround a bug in g++-4.8.3 (and previous versions):
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47679
that is triggered in some cases when boost::optional is used.

Using boost::make_optional to copy-initialize 'bbox' hides the
warning about '*bbox' not being initialized. See that StackOverflow
answer:
  http://stackoverflow.com/a/21755207/1728537

The warning was, for example, with g++-4.8.3:

.../test/Polyhedron_Demo/./CGAL_demo/Scene_interface.h: In member function 'virtual Scene_item::Bbox Scene_polyhedron_selection_item::bbox() const':
.../test/Polyhedron_Demo/./CGAL_demo/Scene_interface.h:32:45: warning: '*((void*)& item_bbox +44)' may be used uninitialized in this function [-Wmaybe-uninitialized]
         xmax(_xmax), ymax(_ymax), zmax(_zmax)
                                             ^
In file included from .../test/Polyhedron_Demo/Scene_polyhedron_selection_item.cpp:1:0:
.../test/Polyhedron_Demo/Scene_polyhedron_selection_item.h:288:35: note: '*((void*)& item_bbox +44)' was declared here
     boost::optional<CGAL::Bbox_3> item_bbox;

That bug is not yet fixed in more recent versions of g++.