Bugfix for Sliding_midpoint splitting rule(default)
Fixed a bug causing the tree to have linear depth in some degenerated
cases, e.g. when points are on a segment or on a hyperplane in general.
Added a check of the tight bounding box to avoid that.

Also set the default bucket_size from 3 to 10, which results in much
better runtime.