change the ways coplanar patches are handled
the previous method was incorrect since some edges inside a
coplanar patch were reported. The following fix let the
intersection of coplanar triangles being computed but
intersection edges are filtered to keep only those
incident to different triangles in both polyhedra.
The drawback of this approach is that the intersection
inside the patch are computed and not really needed.
A better approach would be to be able to directly detect
the boundary of a set of coplanar triangles.