Remove the -g flag in testsuites
While it might seem cleaner to simply add the -g0 argument, removing -g
is safer: -g0 would need to be added after any possible -g flags. Since
CMake makes no guarantees how the COMPILE_OPTIONS of a target are
actually initialized and we cannot change them on an individual basis
this is not easy to achieve. Hence we strip the CXX_FLAGS of -g. This
also still allows specifying it manually through CGAL_CXX_FLAGS.