Definitely removed the "infinite vertex at position 0" constraint...
... and fixed the regular triangulation at the same time.
Even though this constraint had been dropped, the code was still
expecting the infinite vertex to be at position 0 in the cells.
Now, it's not the case anymore, the infinite vertex can be anywhere.
This is particularly useful for the 2D case, because the constraint made
the orientation of the "rightmost" infinite cell wrong.