Check if the libraries are actually defined
In CGALConfig.cmake, the former code was just checking if
WITH_${CGAL_LIB} was TRUE. Now, the new code also checks if the target
was actually defined. If the target is not defined by WITH_${CGAL_LIB}
is ON, if probably means that one dependency of the library was not
properly configured.