Prevent triggering of CMP0054
CMAKE_CXX_COMPILER_ID can expand to the string MSVC, which is also a
variable name which has been expanded by before CMP0054 has been
introduced.

This would lead to a bug *if* we would MATCHES against the string "MSVC",
but only triggers a warning.

Also quote some implicit string in a MATCHES for style reasons.