Always run the config tests
The code tried to prevent running the config tests to often by caching
their values and only rerunning them when the CMAKE_CXX_FLAGS change.

While this is smart it only buys us a few seconds and is not actually
correct. One of those bugs could be triggered through linker flags or
debug flags as well and just checking CMAKE_CXX_FLAGS is far from
enough. To be on the safe side we accept the extra
configuration-time (which is usually also minimized by using compiler
caches).

This also removes a warning caused by CMake policy CMP0054.