Remove unnecessary add_config_flag call
This call to add_config_flag has no effect. It is made before the
file(WRITE ${CMAKE_BINARY_DIR}/include/CGAL/compiler_config.h ...) which
overwrites the contents of the file.

This is surely a bug, because ${CGAL_EXT_LIB_${lib}_PREFIX} expands to
things like GMP or MPFR which would be really bad defines.