attempt to fix the use of "inline"
the warning C4396 happens for these functions on msvc2015

----------------------------------------------
Compiler Warning (level 2) C4396
Visual Studio 2008
"name" : the inline specifier cannot be used when a friend declaration refers to a specialization of a function template

A specialization of a function template cannot specify any of the inline specifiers. The compiler issues warning C4396 and ignores the inline specifier.
To correct this error
    Remove the inline, __inline, or __forceinline specifier from the friend function declaration.
----------------------------------------------