Work around skip functor
We cannot support a skip functor in ray_intersection and we have to work
around the issue of the ray starting inside a facet. To do that we shift
the source of the ray by the epsilon of a floating point number. We have
to consider if we really want to keep it that way.

Add assertions to make sure the new and old code give the same results.