fix a leak in the trapezoidal decomposition that occurs when a curve is removed
A dag node contains a variant that amongst other can point onto a
Tr_active_trapezoid. This class inherits from Handle in order to
reference count the data (of type Tr_active_trapezoid::Data) that
contains 4 neightbors, each being of type Tr_active_trapezoid.
The problem is that if a dag node is no longer used and if it
holds a instance of a Tr_active_trapezoid which neighbors trapezoid
are not cleared, we might have objects of type Tr_active_trapezoid
pointing onto each others preventing them from being deleted.