Udpdate renoir reference platforms for testsuite
During the upgrade of renoir.geometryfactory.com, g++ was upgrade to
g++-5.3.1 (was 4.9.x previously).

That new version of g++ use a dual ABI C++98/C++11. The home-build Boost
versions that were built with g++-4.9.1 introduced issues. I remove
those home-built version of Boost from the config, when the compiler is
g++-5.3.1.

That should fix the recent issues with Boost.Program_options on
renoir.gf.com.