Fix a critical performance issue
As decided by `MainWindow`, the `Scene_c3t3_item::toolTip()` method is
called by `MainWindow::updateInfo()` for each `modified()` event of the
manipulated frame. While the frame is manipulated, that generates a lot
of events, and a lot of calls to `toolTip()`.

Before this commit, the call to `Scene_c3t3_item::toolTip()`
was `O(n)`. After this commit it is `O(1)`.

That speeds up a lot the drawing of the item while the frame is
manipulated!