Improve the draw of the intersection
I have tweaked the vertex and fragment shaders, to avoid the curious
color interpolation for triangles of the intersection with the cutting
plane.

The component `color.w` is set to the signed distance to the cutting
plane (no longer any conditional in the vertex shader).

Then in the fragment shader, the drawing color is composed as such:

    if(color.w<0)
    {
      vec4 my_color = vec4(color.xzy, 1.);
      // [...]
    }
    else
      discard;