Hack to fix a bug with CGAL_HEADER_ONLY
If a `CMakeCache.txt ` is first created without `CGAL_HEADER_ONLY`, and
then `CGAL_HEADER_ONLY` is set, then CMake displays error messages like:
```
CMake Error: Target CGAL has dependency information when it shouldn't.
Your cache is probably stale. Please remove the entry
  CGAL_LIB_DEPENDS
from the cache.
```

The problem comes from the fact that the target `CGAL` was first created as
a library target, and then turned into an "interface library" target.

This patch fixes the error the simplest way: remove the aforementioned
variable from the cache if it is present.