Fix issue #1554
Fix a runtime error. The `Intersect_3` functor of
`Robust_intersection_traits_3_new` was ignored cases in the intersection
between a segment and a triangle.

In `ts_intersection`, if one of the extremities of the segment was
coplanar with the triangle, the empty object was always returned.

This bug of mines was introduced in CGAL-4.3 by
56d733b2546ffa146598901f841567c554f80a65.

(Fix #1554.)