Spatial_searching: Fix strict equality / inconsistent early loop abort.
The method comment suggests the check should be a strict inequality
instead of less-or-equals to test whether the sphere _interior_ intersects.

Additionally, the early abort condition was inconsistent with the final
comparison.
The early abort part is the same kind of bug that was found and fixed in a
related method two weeks ago in this commit:

ed526b8f22ca0d075ca42c6337a39bf5f0421c68
    Spatial_searching: bug reported and fixed by Marc Glisse