Make <CGAL/Polynomial/Cached_extended_euclidean_algorithm.h> thread-safe
The last bug detected by this PR was actually due to a global static
variable. Now that it is a static variable in a function, like others,
the memory de-allocations seem to be done in the right order (even with
Boost implementation of TLS).

Thanks Sébastien @sloriot for the patch.