Fixed calling the VectorH3 constructor using 3FT with underlying negative denominator(s)
If VectorH3 is constructed from 3 FT which are already decomposed in numerators
and denominators, it was possible that the product of the denominators was negative,
which gave an assertion error.

Instead, we use the same mechanism as in the constructor with 4 RT: if the common
denominator would be negative, we flip the signs.