Introduced class Default_instantaneous_mapped_kernel
In KDS, predicates will call rep->static_object() on Point_3 (say, e.g.,
for an orientation predicate). If the triangulation is a regular triangulation,
the points are weighted and are thus in the weighted point table.

Since we don't want to store and maintain the bare points in the point table,
we use this class, which redirects the calls to static_object() and
kinetic_object() that would query the table points to instead query the
weighted points table and return the appropriate bare point.