Fixed depending on a stable random generator to test P3T3
The tests check that the number of elements is the one expected, but if the
random generator changes (even for a fixed seed), this breaks everything...

Instead, we dump the random generator for a given seed and use this output
as input every time.

This commit is inspired by the playstation network.