Fix compilation errors with new implement of assert() in glibc
In glibc 2.25 and 2.26, the implementation of `assert()` triggers
compilation errors with values that are not comparable to an `int`, such
a iostream, or `optional`.

That is the "bug" https://sourceware.org/bugzilla/show_bug.cgi?id=21972

In my opinion that is not really a bug: the value test by `assert` is
supposed to be an integral value, and not something explicitly
convertible to `bool`.

This commit uses an explicit conversion to `bool`, in all problematic
calls to `assert()`.

The problem was seen on my platform, Linux Fedora 26.