Added an assertion to forbid using ExactComparisonTag with periodic triangulations
When we set the exact comparison tag to true, we use a lazy evaluation
of predicates and store pointers to the arguments of the predicates (that is,
pointer to points). However, the points are  -- in the case of periodic --
only temporary objects and it is thus dangerous to take pointers to those
temporary values because the pointers quickly become invalid.

Thus, periodic triangulations are not allowed to use the exact tag.

A possible way to make it work is to define a small class:
Periodic_triangulation_with_stored_real_points that would inherit the base
(periodic) triangulation but store all the points of its simplices.
Then, the function  point(face_handle, int) would return an entry of that
container, of which it would be safe to take a pointer.