Cherry-pick pull request #2409 from SKURInc/master
avoid index out of range with large bitmap sizes

With large u and v, the product u * v exceeds the capacity of an int.  The result when coerced to size_t can be an out of range index into bitmap.
Avoid this by casting u and v to size_t before multiplying.