Remote that test...
Long time ago, CGAL was using the Boost Result_of library, that prevailed
before `std::tr1::result_of` and then `std::result_of` were standardized in
the C++ norm. Then CGAL made a switch to using the wrapper
`CGAL::cpp11::result_of`, implemented either with `tr1::result_of`, or a
C++11 decltype-based implementation...

The test is just checking that, when C++11 the decltype implementation is
used, the result is the same as with `tr1::result_of`... Now that we know
it works, let's remove the test (that no longer compile with Boost 1.65).